webapp/requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==3.12.0
django-model-utils==4.2.0
djangorestframework==3.13.1
drf-writable-nested==0.6.3
install==1.3.5
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
